requirements.txt

Django>=5.0.0,<5.1
djangorestframework>=3.14.0,<3.15
psycopg2==2.9.9
drf-spectacular>=0.27.2,<0.28
